Add more authorization

+++ api/Appointment/delete.php
@@ -1,4 +1,6 @@
 <?php
+  session_start();
+  
   header('Access-Control-Allow-Origin: *');
   header('Content-Type: application/json');
   header('Access-Control-Allow-Methods: DELETE');
@@ -7,19 +9,27 @@
   include_once '../../config/Database.php';
   include_once '../../model/Appointment.php';
 
-  $database = new Database();
-  $db = $database->connect();
-	$ap = new Appointment($db);
-  $data = json_decode(file_get_contents("php://input"));
+  if (isset($_SESSION['user_id'])
+    && isset($_SESSION['user_type'])
+    && $_SESSION['user_type'] == 'admin'
+  ) {
+    $database = new Database();
+    $db = $database->connect();
+    $ap = new Appointment($db);
+    $data = json_decode(file_get_contents("php://input"));
 
-  $ap->appoint_id = $data->appoint_id;
+    $ap->appoint_id = $data->appoint_id;
 
-  if ($ap->delete()) {
-    echo json_encode(
-      array('message' => 'Appointment Deleted')
-    );
+    if ($ap->delete()) {
+      echo json_encode(
+        array('message' => 'Appointment Deleted')
+      );
+    } else {
+      echo json_encode(
+        array('message' => 'Appointment Not Deleted')
+      );
+    }
   } else {
-    echo json_encode(
-      array('message' => 'Appointment Not Deleted')
-    );
+    http_response_code(401);
+    echo json_encode(array('message' => 'You Are Not Authorized'));
   }
