+++ api/Calendar/delete.php
@@ -1,4 +1,6 @@
-<?php 
+<?php
+  session_start();
+  
   header('Access-Control-Allow-Origin: *');
   header('Content-Type: application/json');
   header('Access-Control-Allow-Methods: DELETE');
@@ -7,20 +9,28 @@
   include_once '../../config/Database.php';
   include_once '../../model/Calendar.php';
 
-  $database = new Database();
-  $db = $database->connect();
-	$calendar = new Calendar($db);
-  $data = json_decode(file_get_contents("php://input"));
-  
-  $calendar->month = $data->month;
-  $calendar->year = $data->year;
+  if (isset($_SESSION['user_id'])
+    && isset($_SESSION['user_type'])
+    && $_SESSION['user_type'] == 'admin'
+  ) {
+    $database = new Database();
+    $db = $database->connect();
+    $calendar = new Calendar($db);
+    $data = json_decode(file_get_contents("php://input"));
+    
+    $calendar->month = $data->month;
+    $calendar->year = $data->year;
 
-  if ($calendar->delete()) {
-    echo json_encode(
-      array('message' => 'Calendar Deleted')
-    );
+    if ($calendar->delete()) {
+      echo json_encode(
+        array('message' => 'Calendar Deleted')
+      );
+    } else {
+      echo json_encode(
+        array('message' => 'Calendar Not Deleted')
+      );
+    }
   } else {
-    echo json_encode(
-      array('message' => 'Calendar Not Deleted')
-    );
+    http_response_code(401);
+    echo json_encode(array('message' => 'You Are Not Authorized'));
   }
