+++ api/Employee/delete.php
@@ -1,4 +1,6 @@
 <?php
+  session_start();
+
   header('Access-Control-Allow-Origin: *');
   header('Content-Type: application/json');
   header('Access-Control-Allow-Methods: DELETE');
@@ -7,19 +9,27 @@
   include_once '../../config/Database.php';
   include_once '../../model/Employee.php';
 
-  $database = new Database();
-  $db = $database->connect();
+  if (isset($_SESSION['user_id'])
+    && isset($_SESSION['user_type'])
+    && $_SESSION['user_type'] == 'admin'
+  ) {
+    $database = new Database();
+    $db = $database->connect();
 
-  $employee = new Employee($db);
-  $data = json_decode(file_get_contents("php://input"));
-  $employee->user_id = $data->user_id;
+    $employee = new Employee($db);
+    $data = json_decode(file_get_contents("php://input"));
+    $employee->user_id = $data->user_id;
 
-  if($employee->delete()) {
-    echo json_encode(
-      array('message' => 'Employee Deleted')
-    );
+    if($employee->delete()) {
+      echo json_encode(
+        array('message' => 'Employee Deleted')
+      );
+    } else {
+      echo json_encode(
+        array('message' => 'Employee Not Deleted')
+      );
+    }
   } else {
-    echo json_encode(
-      array('message' => 'Employee Not Deleted')
-    );
+    http_response_code(401);
+    echo json_encode(array('message' => 'You Are Not Authorized'));
   }
