+++ api/HealthReport/delete.php
@@ -1,4 +1,6 @@
 <?php
+  session_start();
+
   header('Access-Control-Allow-Origin: *');
   header('Content-Type: application/json');
   header('Access-Control-Allow-Methods: DELETE');
@@ -7,20 +9,28 @@
   include_once '../../config/Database.php';
   include_once '../../model/HealthReport.php';
 
-  $database = new Database();
-  $db = $database->connect();
-	$hp = new HealthReport($db);
-  $data = json_decode(file_get_contents("php://input"));
+  if (isset($_SESSION['user_id'])
+    && isset($_SESSION['user_type'])
+    && $_SESSION['user_type'] == 'admin'
+  ) {
+    $database = new Database();
+    $db = $database->connect();
+    $hp = new HealthReport($db);
+    $data = json_decode(file_get_contents("php://input"));
 
-  $hp->client_id = $data->client_id;
-  $hp->date = $data->date;
+    $hp->client_id = $data->client_id;
+    $hp->date = $data->date;
 
-  if ($hp->delete()) {
-    echo json_encode(
-      array('message' => 'Health Report Deleted')
-    );
+    if ($hp->delete()) {
+      echo json_encode(
+        array('message' => 'Health Report Deleted')
+      );
+    } else {
+      echo json_encode(
+        array('message' => 'Health Report Not Deleted')
+      );
+    }
   } else {
-    echo json_encode(
-      array('message' => 'Health Report Not Deleted')
-    );
+    http_response_code(401);
+    echo json_encode(array('message' => 'You Are Not Authorized'));
   }
